Fix label->checkbox assignment
Previously, all the text for all options was incorrectly marked as the
label of the first checkbox. As a result, clicking on any of the text
would *confusingly* toggle the 5.1 audio option. This commit assigns
each label to an individual checkbox -- each checkbox can be toggled by
clicking on its respective option.